requirements.txt
aiohttp
aiofiles
aioshutil
anytree
apscheduler
aria2p
asyncio
cloudscraper
dnspython
feedparser
flask
gevent
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gunicorn
httpx
lxml
myjdapi
motor
natsort
pillow
psutil
pycryptodome
langcodes
pymongo
pyrofork==2.2.11
python-dotenv
pytz
bs4
python-magic
gtts
qbittorrent-api
requests
telegraph
tenacity
waybackpy
tgcrypto
urllib3
uvloop
xattr
speedtest-cli
yt-dlp[default,curl-cffi]
heroku3
html_telegraph_poster
